+++ HISTORY.md
@@ -14,6 +14,7 @@
 * updated Sprite Method API library for the new slot event hat block
 * updated German translation: removed obsolete strings of deprecated slot-interaction blocks
 * updated German translation: added new strings for slot-interaction blocks
+* objects: changed default value for slot-event to "menu"
 
 ### 2024-11-09
 * threads: enhanced "set slot ... to ..." command to also handle variadic inputs, setting them to a list of values